requirements/dev.txt
-r ./common.txt

django-debug-toolbar==2.2

#testing.
coverage==4.0.3
webtest==2.0.9
django-webtest==1.7.9
model_mommy==1.2.6
mock==2.0.0

#interactive interpreter for debugging
ipython
